+++ src/main/java/org/truffleruby/parser/YARPMultiWriteNodeTranslator.java
@@ -20,6 +20,9 @@ import org.truffleruby.core.cast.SplatCastNode;
 import org.truffleruby.core.cast.SplatCastNodeGen;
 import org.truffleruby.language.RubyNode;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /** Translate Nodes.MultiWriteNode node.
  *
  * NOTE: cannot inherit from YARPBaseTranslator because it returns AssignableNode instead of RubyNode. */
@@ -28,6 +31,10 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
     private final Nodes.MultiWriteNode node;
     private final RubyLanguage language;
     private final YARPTranslator yarpTranslator;
+    /** Nodes to initialize local variables before assign values. Store in the local variables receivers of attribute
+     * (a.b = ...) and reference (a[b] = ...) assignments as well as fully qualified constants' modules (A::B = ...).
+     * They should be evaluated before values to conform with the left-to-right order semantic. */
+    private final List<RubyNode> prolog;
 
     public YARPMultiWriteNodeTranslator(
             Nodes.MultiWriteNode node,
@@ -36,10 +43,10 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
         this.node = node;
         this.language = language;
         this.yarpTranslator = yarpTranslator;
+        this.prolog = new ArrayList<>();
     }
 
     public RubyNode translate() {
-        final RubyNode rubyNode;
         final RubyNode rhsNode = node.value.accept(yarpTranslator);
         final SplatCastNode splatCastNode = SplatCastNodeGen.create(
                 language,
@@ -70,13 +77,13 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
             postNodes[i] = node.rights[i].accept(this);
         }
 
-        rubyNode = new MultipleAssignmentNode(
+        return new MultipleAssignmentNode(
+                prolog.toArray(RubyNode.EMPTY_ARRAY),
                 preNodes,
                 restNode,
                 postNodes,
                 splatCastNode,
                 rhsNode);
-        return rubyNode;
     }
 
     @Override
@@ -87,12 +94,24 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
 
     @Override
     public AssignableNode visitCallTargetNode(Nodes.CallTargetNode node) {
+        // store receiver in a local variable to evaluate before assigned value
+        Nodes.Node readReceiver = stash(node.receiver, "receiver");
+
+        node = new Nodes.CallTargetNode(node.flags, readReceiver, node.name, node.startOffset, node.length);
+
         final RubyNode rubyNode = node.accept(yarpTranslator);
         return ((AssignableNode) rubyNode).toAssignableNode();
     }
 
     @Override
     public AssignableNode visitConstantPathTargetNode(Nodes.ConstantPathTargetNode node) {
+        if (node.parent != null) {
+            // store parent lexical scope (e.g foo in foo::C = ...) in a local variable to evaluate before assigned value
+            Nodes.Node readParent = stash(node.parent, "parent");
+
+            node = new Nodes.ConstantPathTargetNode(readParent, node.child, node.startOffset, node.length);
+        }
+
         final RubyNode rubyNode = node.accept(yarpTranslator);
         return ((AssignableNode) rubyNode).toAssignableNode();
     }
@@ -116,6 +135,27 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
 
     @Override
     public AssignableNode visitIndexTargetNode(Nodes.IndexTargetNode node) {
+        // store receiver in a local variable to evaluate before assigned value
+        Nodes.Node readReceiver = stash(node.receiver, "receiver");
+
+        // store arguments in local variables to evaluate after receiver but before assigned values
+        final Nodes.ArgumentsNode arguments;
+        if (node.arguments != null) {
+            var argumentsReads = new Nodes.Node[node.arguments.arguments.length];
+
+            for (int i = 0; i < node.arguments.arguments.length; i++) {
+                argumentsReads[i] = stash(node.arguments.arguments[i], "argument");
+            }
+
+            arguments = new Nodes.ArgumentsNode(node.arguments.flags, argumentsReads, node.arguments.startOffset,
+                    node.arguments.length);
+        } else {
+            arguments = null;
+        }
+
+        node = new Nodes.IndexTargetNode(node.flags, readReceiver, arguments, node.block, node.startOffset,
+                node.length);
+
         final RubyNode rubyNode = node.accept(yarpTranslator);
         return ((AssignableNode) rubyNode).toAssignableNode();
     }
@@ -137,6 +177,8 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
         final var translator = new YARPMultiTargetNodeTranslator(node, language, yarpTranslator, null);
         final MultipleAssignmentNode multipleAssignmentNode = translator.translate();
 
+        prolog.addAll(translator.prolog);
+
         return multipleAssignmentNode.toAssignableNode();
     }
 
@@ -153,6 +195,18 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
     @Override
     protected AssignableNode defaultVisit(Nodes.Node node) {
         throw new Error("Unknown node: " + node);
+    }
+
+    /** Cache node evaluation result in a local variable to execute before assigning */
+    Nodes.Node stash(Nodes.Node node, String name) {
+        var e = new YARPExecutedOnceExpression(name, node, yarpTranslator);
+        RubyNode writeNode = e.getWriteNode();
+
+        if (writeNode != null) {
+            prolog.add(writeNode);
+        }
+
+        return e.getReadYARPNode();
     }
 
 }